refacto: split sections in hompage into separate components in Homepage folder
